noor_chatbot/requirements.txt
actionlib==1.14.0
angles==1.9.13
annotated-types==0.7.0
base_local_planner==1.17.3
bondpy==1.8.6
cachetools==5.5.2
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
defusedxml==0.7.1
diagnostic-updater==1.11.0
distro==1.9.0
docutils==0.21.2
dynamic-reconfigure==1.7.3
empy==4.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-speech==2.32.0
google-cloud-texttospeech==2.27.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
image-geometry==1.16.2
interactive-markers==1.12.0
joint-state-publisher==1.15.1
laser_geometry==1.6.7
message-filters==1.16.0
pkg_resources==0.0.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.11.5
pydantic_core==2.33.2
pygame==2.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
PyYAML==6.0.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-app==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.32.3
resource_retriever==1.12.7
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.6.0
rospy==1.16.0
rosserial_python==0.9.2
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rsa==4.9.1
rviz==1.14.20
sensor-msgs==1.13.1
six==1.17.0
smclib==1.8.6
tf==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
topic-tools==1.16.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
xacro==1.14.18
